lib/nvme: get nsid, logical block size and size for a generic device.
This is a prep patch. It modifies the fio_file struct to store nsid and
logical_block_size. Also a get_char_size handler is used to get the size
of generic device. This will be needed to enable runtime option when
using fio for doing I/O on a generic device.

Signed-off-by: Anuj Gupta <anuj20.g@samsung.com>